feat: add training history controller
