+++ lib/screens/progress_overview_screen.dart
@@ -4,8 +4,8 @@ import 'package:provider/provider.dart';
 import '../services/progress_forecast_service.dart';
 import '../models/training_result.dart';
 import '../widgets/ev_icm_trend_chart.dart';
-import '../widgets/common/accuracy_chart.dart';
-import '../widgets/common/average_accuracy_chart.dart';
+import '../widgets/history/accuracy_chart.dart';
+import '../widgets/history/average_accuracy_chart.dart';
 import '../widgets/common/ev_icm_trend_chart.dart' as common;
 import '../widgets/sync_status_widget.dart';
 import '../theme/app_colors.dart';
